% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawASEPlots.R
\name{drawBoxPlot}
\alias{drawBoxPlot}
\title{Draw boxplot}
\usage{
drawBoxPlot(out, nam, minExp = 5)
}
\arguments{
\item{out}{ASE expression info}

\item{nam}{Name of snp/gene pair you want to plot}

\item{minExp}{Min number of phased UMIs to plot}
}
\value{
A plot!
}
\description{
Draw boxplot
}
