% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTransEQTL.R
\name{GetGeneLevel}
\alias{GetGeneLevel}
\title{Get gene level trans-eQTL analysis}
\usage{
GetGeneLevel(mrk)
}
\arguments{
\item{mrk}{A data frame generated by combining the results from GetTransEQTL for multiple SNP/genes}
}
\value{
A table with pvalues and FDR at the genes level
}
\description{
Combines the results of the trans-eQTL analysis to see if there are genes that are overrepresented among the top hits.
}
