% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadSNPs.R
\name{GetSNPs}
\alias{GetSNPs}
\title{Load SNP Data}
\usage{
GetSNPs(
  seur = NULL,
  meta = NULL,
  genes,
  snp_list = c(),
  samp = "orig.ident",
  snp_col = "SNP",
  all_col = "Allele",
  cbc_col = "",
  cond = "PD",
  countCells = F,
  cellType = "",
  bulk = T
)
}
\arguments{
\item{seur}{A seurat object prepared to have the correct columns (see below). Assumes the cell names in Seurat are \link{samp name}_\link{CBC} or CBC is given as column}

\item{meta}{A dataframe with meta data, can be given instead of Seurat object}

\item{genes}{A list of genes to get SNP specific information for}

\item{snp_list}{List of SNPs, if given will only return snps from the list, must be the same length as genes}

\item{snp_col}{The name of a column in seur@meta.datapointing to the SNPLevelCounts/snps.bed output from the AI pipeline for this cell}

\item{all_col}{The name of a column in seur@meta.datapointing to the AlleleCounts/counts.txt output from the AI pipeline for this cell}

\item{cbc_col}{Location of cell barcode in meta data, if not given assums sample names are \link{samp name}_\link{CBC}}

\item{cond}{The column in seur@meta.data with condition ifnormation}

\item{countCells}{If True returns cell counts instead of read counts}

\item{cellType}{The name of the column in seur@meta.data with cell type information if required}

\item{bulk}{If true makes pseudobulk, else returns at cell level}
}
\value{
A data frame with bulk allele expression information at the SNP level
}
\description{
Loads SNP data for genes of interest.
}
