% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTransEQTL.R
\name{GetTransEQTL}
\alias{GetTransEQTL}
\title{Get trans eqtl}
\usage{
GetTransEQTL(
  meta,
  gene,
  snp,
  tab = NULL,
  dat = NULL,
  samp_col = "orig.ident",
  snp_col = "SNP",
  raw_col = "Raw",
  all_col = "Allele",
  cbc_col = "CBC",
  cond = "PD",
  on_disk = T,
  method = "nb"
)
}
\arguments{
\item{meta}{The meta data as produced by LoadPipeline}

\item{gene}{The gene whose ASE is being tested}

\item{snp}{The snp used to align alleles between samples}

\item{tab}{The SNP level allele specific information, calculated internally if not given}

\item{dat}{The UMI expression data, calculated internally if not given}
}
\value{
A table with the results of the trans-eqtl testing
}
\description{
A tool that uses variation in ASE between cells to help estimate trans-eQTL for a given gene/SNP pair.
}
