% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadAIGene.R
\name{loadASEGene}
\alias{loadASEGene}
\title{Load Gene ASE Data}
\usage{
loadASEGene(
  seur = NULL,
  meta = NULL,
  samp = "orig.ident",
  all_col = "Allele",
  cbc_col = "CBC",
  cond = "PD",
  celltype = "CellType"
)
}
\arguments{
\item{seur}{A seurat object prepared to have the correct columns (see below). Assumes the cell names in Seurat are \link{samp name}_\link{CBC} or CBC is given as column}

\item{meta}{A dataframe with meta data, can be given instead of Seurat object}

\item{all_col}{The name of a column in seur@meta.datapointing to the AlleleCounts/counts.txt output from the AI pipeline for this cell}

\item{cbc_col}{Location of cell barcode in meta data, if not given assums sample names are \link{samp name}_\link{CBC}}

\item{cond}{The column in seur@meta.data with condition ifnormation}

\item{cellType}{The name of the column in seur@meta.data with cell type information if required}
}
\value{
A data frame with allele expression information at the gene level
}
\description{
Loads Gene level ASE data
}
