% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadExpression.R
\name{LoadCountMatrix}
\alias{LoadCountMatrix}
\title{Load expression}
\usage{
LoadCountMatrix(dirRaw)
}
\arguments{
\item{dirRaw}{The directory with expression info}
}
\value{
Sparse count matrix
}
\description{
Loads MM data as produced by STARSolo. Gene names are the gene name and gene id field combined.
}
