% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadPipeNice.R
\name{LoadData}
\alias{LoadData}
\title{Loads data from the pipeline as used in practice}
\usage{
LoadData(
  fils,
  var_gene_only = T,
  numVar = 10,
  minGenes = 500,
  regress = "nFeature_RNA"
)
}
\arguments{
\item{fils}{List of directories}

\item{var_gene_only}{If true only load variable genes}

\item{numVar}{Number of samples a gene must be variable in to be used}

\item{minGenes}{Min number of genes}

\item{regress}{Covariates to regress out}
}
\value{
A Seurat object
}
\description{
Loads the data into a Seurat object
}
