% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadExpression.R
\name{LoadExpression}
\alias{LoadExpression}
\title{Loads expression data based on meta data}
\usage{
LoadExpression(
  meta,
  cbc_col = "CBC",
  raw_col = "Raw",
  samp_col = "orig.ident",
  genes_include = NULL
)
}
\arguments{
\item{meta}{The metadata with pipeline info as produced by LoadMeta}

\item{cbc_col}{Name of column with cell barcode info}

\item{raw_col}{Which column contains raw count data}

\item{samp_col}{Name of the column with sample of origin info}
}
\value{
A sparse count matrix corresponding to meta
}
\description{
Load the expression data from the pipeline.
}
