% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadPipelineOutput.R
\name{loadIntoMatrices}
\alias{loadIntoMatrices}
\title{Load Pipeline Into Matrix}
\usage{
loadIntoMatrices(dirs)
}
\arguments{
\item{dirs}{A named list of directories from the AI pipeline}
}
\value{
A list of matrices, one with Gene level Expression (named Expression) and one with allele level expression for each gene (names Allele)
}
\description{
Loads the results of the pipeline into a list of matrices
}
