% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadMeta.R
\name{LoadPipeline}
\alias{LoadPipeline}
\title{Add info about pipeline output to meta data}
\usage{
LoadPipeline(meta, lst, samp_col = "orig.ident", cbc_col = "CBC")
}
\arguments{
\item{meta}{A data frame with cell level info}

\item{lst}{An named character array, one entry for each sample, named by sample (should agree with the naming in meta)}

\item{samp_col}{The name of the column in meta with info about sample of origin}

\item{cbc_col}{The name of the column in meta with info about cbc (can be excluded for now)}
}
\value{
The meta dataframe with the added columns about the pipeline added
}
\description{
Given a data frame containing meta data on a per cell basis and names list mapping sample to pipeline output directory, adds the pipeline info to the meta data
}
