% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadSNPs.R
\name{loadSNPs2}
\alias{loadSNPs2}
\title{Load SNP data}
\usage{
loadSNPs2(
  samp,
  snp,
  allele,
  nams,
  genes,
  celltypes,
  getVals = F,
  countCells = F,
  snp_list = c(),
  bulk = T
)
}
\arguments{
\item{samp}{Sample name}

\item{snp}{Location of snps.bed file}

\item{allele}{Gene level ASE count data}

\item{nams}{Cell names to use}

\item{genes}{List of genes to load SNP data for}

\item{celltypes}{Cell type information}

\item{getVals}{For testing}

\item{countCells}{For testing}

\item{snp_list}{A list of SNPs to look at, must be the same length as genes}

\item{bulk}{If true makes pseudobulk, else returns at cell level}
}
\value{
A data from with SNP level ASE information
}
\description{
Helper function to load SNP level ASE data (new version)
}
