% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadPipelineOutput.R
\name{MakeSeurat}
\alias{MakeSeurat}
\title{Make Seurat object}
\usage{
MakeSeurat(dat_gene, dat_allele, minGenes = 500, regress = "nFeature_RNA")
}
\arguments{
\item{dat_gene}{A matrix of gene expression, cells in the columns, genes in the rows}

\item{dat_allele}{A matrix of allele specific information, cells in the columns, alleles in the rows}

\item{minGenes}{The min number of genes below which a cell will be removed}

\item{regress}{The parameters to be regressed out in the ScaleData step}
}
\value{
A Seurat object with gene level data in the RNA assay, Allele specific data in the ALLELE assay
}
\description{
Makes a Seurat object with gene level expression and allele specific expression. Meant to work with the loadIntoMatrices function below.
}
