% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadPipeNice.R
\name{NormData}
\alias{NormData}
\title{Normalize data}
\usage{
NormData(seur)
}
\arguments{
\item{seur}{The seurat object}
}
\value{
Seurat object with normalization
}
\description{
An alternative to the NormalizeData from Seurat
}
