% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadPipelineOutput.R
\name{ReadInAllele}
\alias{ReadInAllele}
\title{Read in Gene level allele information}
\usage{
ReadInAllele(fils)
}
\arguments{
\item{fils}{A list of files with gene level allele specific expression data (in the AlleleCounts/counts.txt subdirectory produced by AI pipeline)}
}
\value{
A sparse matrix of gene level allele specific expression data
}
\description{
Given a list of allele level count data from the AI pipeline reads the results into a matrix
}
