% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadPipelineOutput.R
\name{ReadInSNPs}
\alias{ReadInSNPs}
\title{Read in table of SNP info}
\usage{
ReadInSNPs(fils)
}
\arguments{
\item{fils}{A list of comb.bed files to load in}
}
\value{
Returns a matric with 4 columns corresponding to the SNP, the gene, the Allele from the vcf (0|1 or 1|0) telling you if which value is in the ref vs alt, and the Sample name
}
\description{
Takes in the comb.bed SNP file produced by the ASE Pipeline and returns as a data.frame
}
