% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadPipelineOutput.R
\name{ReadInSTARSolo}
\alias{ReadInSTARSolo}
\title{Read in STARSolo data}
\usage{
ReadInSTARSolo(
  dirs,
  runFilts = F,
  numCells = 10000,
  var_genes = c(),
  numCellsSamp = 0
)
}
\arguments{
\item{dirs}{A (possibly named) list of directories with STARSolo output}

\item{runFilts}{If T performs filtering with EmptyDroplets}

\item{numCells}{If filtering, numCells is the expected number of cells}

\item{var_genes}{A list of variable genes, if non-empty will only return the data for those genes}

\item{numCellSamp}{If >0 subsamples to the given number of cells}
}
\value{
Expression information in sparse matrix form
}
\description{
Reads data out of STARSolo and into a matrix
}
