% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunDAESC.R
\name{RunDAESC}
\alias{RunDAESC}
\title{Run DAESC}
\usage{
RunDAESC(
  dat,
  minCells = 50,
  minSamps = 10,
  form = ~Condition,
  formNull = ~1,
  method = "bb"
)
}
\arguments{
\item{dat}{The ASE count data produced by the GetSNPs command}

\item{minCells}{The min number of cells with at least one phased UMI in a SNP required to test a SNP}

\item{minSamps}{The min number of samples with at least one phased UMI in a SNP required to test a SNP}

\item{form}{The formula for the model to use in LRT}

\item{formNull}{The formula for the null model to compare to}

\item{method}{The name of the DAESC method to use, either bb or mix (if anything else is passed will use mix)}
}
\value{
A data frame with the results for the DAESC method
}
\description{
A method that takes in the count data produced by this package and uses DAESC with it.
}
