% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAI_betabin.R
\name{TestSNP_betabin}
\alias{TestSNP_betabin}
\title{Test SNP for allelic imbalance with BetaBinomial Regression}
\usage{
TestSNP_betabin(
  dat,
  minCount = 50,
  minSamp = 10,
  form = cbind(alt, Num - alt) ~ 1,
  form2 = cbind(alt, Num - alt) ~ 0,
  coef_ret = "(Intercept)",
  numTest = -1
)
}
\arguments{
\item{dat}{A dataframe with columns corresponding to the expression of each SNP. Should be a column called Feature (the SNP of interest), Gene (the gene being tested), Sample (which individual it came from), and Allele (alt or ref). The naive model just tests for presence of allelic imbalance at the SNP, but can test for conditional as well}

\item{minCount}{The min number of reads over all samples for the SNP to be tested}

\item{minSamp}{The min number of samples with allelic information about this gene for it to be used}

\item{form}{The formula for the betabinomial model being fitted. The default works in most cases in a format as used by aod}

\item{coef_ret}{The coefficient to be tested (in most cases the intercept)}

\item{numTest}{Used mostly for debugging, will only test this many SNPs and ignore the rest}
}
\value{
A dataframe with the results of the allelic imbalance test
}
\description{
Combines allele expression data for numerous individuals to see if there is significant allelic imbalance at that SNP.
Based on a betabinomial model
}
