% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAI_nb_LRT.R
\name{TestSNP_LRT}
\alias{TestSNP_LRT}
\title{Test SNP for allelic imbalance with Negative Binomial Regression LRT}
\usage{
TestSNP_LRT(
  dat,
  minCount = 50,
  minSamp = 10,
  form = Count ~ Allele + Sample,
  coef_ret = "Allele",
  numTest = -1
)
}
\arguments{
\item{dat}{A dataframe with columns corresponding to the expression of each SNP. Should be a column called Feature (the SNP of interest), Gene (the gene being tested), Sample (which individual it came from), and Allele (alt or ref). The naive model just tests for presence of allelic imbalance at the SNP, but can test for conditional as well (for example using form=Count~Allele+Sample+Condition*Allele,coef_ret="Alleleref:Conditionyes")}

\item{minCount}{The min number of reads over all samples for the SNP to be tested}

\item{minSamp}{The min number of samples with allelic information about this gene for it to be used}

\item{form}{The formula for the negative binomial model being fitted. The default works in most cases}

\item{coef_ret}{The coefficient to be tested (in most cases Allele)}

\item{numTest}{Used mostly for debugging, will only test this many SNPs and ignore the rest}
}
\value{
A dataframe with the results of the allelic imbalance test
}
\description{
Combines allele expression data for numerous individuals to see if there is significant allelic imbalance at that SNP.
Based on a negative binomial model with a LRT test
}
