% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTransEQTL.R
\name{TestTransEQTL}
\alias{TestTransEQTL}
\title{Test trans-eQTL}
\usage{
TestTransEQTL(
  meta,
  dat,
  samp_col = "orig.ident",
  pseudo = 1,
  percCut = 0.05,
  on_disk = T,
  method = "nb"
)
}
\arguments{
\item{meta}{Meta data frame including ASE data}

\item{dat}{Matrix (can be sparse) of expression counts, genes by cells}

\item{samp_col}{The column with sample of origin}

\item{pseudo}{Pseudocount to use (corresponds to prior in beta prior)}

\item{percCut}{Percent of cells that need to express a gene for it to be tested}
}
\description{
Given ASE data and count data tests for transeQTL
}
